Remove support for HTML-compatibility `list-style-type` values in CSS
These are neither in the specs, nor supported by current browsers. The
values are mapped to regular counter-style names in the
`AttributeTranslator` class now.

To clarify: `type="A"` is still supported in HTML, and translated to
`list-style-type: upper-alpha;`, but support for `list-style-type: A;`
is removed (treated as `list-style-type: decimal;` now because
the name is unrecognized).

https://www.w3.org/TR/CSS2/generate.html#list-style
https://www.w3.org/TR/css-counter-styles-3/#predefined-counters